+++ rye/src/bootstrap.rs
@@ -31,11 +31,11 @@ pub const SELF_PYTHON_TARGET_VERSION: PythonVersionRequest = PythonVersionReques
     suffix: None,
 };
 
-const SELF_VERSION: u64 = 16;
+const SELF_VERSION: u64 = 17;
 
 const SELF_REQUIREMENTS: &str = r#"
-build==1.0.3
-certifi==2023.11.17
+build==1.1.1
+certifi==2024.2.2
 charset-normalizer==3.3.2
 click==8.1.7
 distlib==0.3.8